# openapi.yaml
openapi: 3.0.0
info:
  title: Workflow Approval System API
  description: API for a SaaS platform that allows customers to manage approvals for generic workflows, users, and groups.
  version: 1.0.0
servers:
  - url: https://approvio.com/api/v1
    description: Production server
paths:
...
  /users:
    post:
      summary: Create a new user
      description: Creates a new user in the system.
      operationId: createUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: User with this email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
    get:
      summary: List users
      description: Returns a list of users, optionally filtered by group ID.
      operationId: listUsers
      tags:
        - Users
      parameters:
        - name: groupId
          in: query
          description: Filter users by the group they belong to.
          required: false
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of results per page
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/{userId}:
    get:
      summary: Get user details
      description: Returns the details of a specific user by their ID.
      operationId: getUser
      tags:
        - Users
      parameters:
        - name: userId
          in: path
          required: true
          description: The unique identifier (UUID) of the user.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /groups:
    post:
      summary: Create a new approver group
      description: Create a new group for organizing approvers
      operationId: createGroup
      tags:
        - Groups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupCreate"
      responses:
        "201":
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      summary: List groups
      description: Returns a list of groups
      operationId: listGroups
      tags:
        - Groups
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of results per page
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: object
                properties:
                  groups:
                    type: array
                    items:
                      $ref: "#/components/schemas/Group"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
                required:
                  - groups
                  - pagination
        "401":
          $ref: "#/components/responses/Unauthorized"
  /groups/{groupIdentifier}:
    get:
      summary: Get group details
      description: Returns the details of a specific group
      operationId: getGroup
      tags:
        - Groups
      parameters:
        - name: groupIdentifier
          in: path
          required: true
          description: The unique identifier of the group or the name of the group
          schema:
            type: string
      responses:
        "200":
          description: Group details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /groups/{groupId}/entities:
    get:
      summary: List entities in a group
      description: Returns a list of all entities (users, systems, etc.) that belong to a specific group, including their roles.
      operationId: listGroupEntities
      tags:
        - Groups
        - Users
      parameters:
        - name: groupId
          in: path
          required: true
          description: The unique identifier (UUID) of the group.
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of results per page
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of entities in the group
          content:
            application/json:
              schema:
                type: object
                required:
                  - entities
                  - pagination
                properties:
                  entities:
                    type: array
                    items:
                      $ref: "#/components/schemas/GroupMembership"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      summary: Add group entities
      description: Add entities (like users) within a specific group.
      operationId: addGroupEntities
      tags:
        - Groups
        - Users
      parameters:
        - name: groupId
          in: path
          required: true
          description: The unique identifier (UUID) of the group.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddGroupEntitiesRequest"
      responses:
        "200":
          description: Entities managed successfully. Returns updated group details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "409":
          description: Conflict detected (e.g., adding an existing member).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
    delete:
      summary: Remove group entities
      description: Remove entities (like users) from a specific group.
      operationId: removeGroupEntities
      tags:
        - Groups
        - Users
      parameters:
        - name: groupId
          in: path
          required: true
          description: The unique identifier (UUID) of the group.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemoveGroupEntitiesRequest"
      responses:
        "200":
          description: Entities removed successfully. Returns updated group details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "409":
          description: Conflict detected (e.g. removing a non-member).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
components:
  schemas:
    UserCreate:
      type: object
      required:
        - displayName
        - email
        - orgRole
      properties:
        displayName:
          type: string
          description: User's display name.
          example: "Alex Chen"
        email:
          type: string
          format: email
          description: User's email address (must be unique).
          example: "alex.chen@example.com"
        orgRole:
          type: string
          x-extensible-enum:
            - admin
            - member
          description: Role assigned to the user within the group.
          example: "member"
    User:
      type: object
      required:
        - id
        - displayName
        - email
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Internal unique identifier for the user.
          readOnly: true
          example: "a1b2c3d4-e5f6-7890-1234-567890abcdef"
        displayName:
          type: string
          description: User's display name.
          example: "Alex Chen"
        email:
          type: string
          format: email
          description: User's email address.
          example: "alex.chen@example.com"
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the user was created.
          readOnly: true
          example: "2025-04-15T12:05:00Z"
    EntityReference:
      type: object
      required:
        - entityType
        - entityId
      properties:
        entityType:
          type: string
          x-extensible-enum:
            - human
            - system
          description: The type of the entity being referenced.
          example: human
        entityId:
          type: string
          format: uuid
          description: The unique identifier (UUID) of the entity.
          example: "a1b2c3d4-e5f6-7890-1234-567890abcdef"
    EntityMembershipAdd:
      type: object
      required:
        - entity
        - role
      properties:
        entity:
          $ref: "#/components/schemas/EntityReference"
        role:
          type: string
          x-extensible-enum:
            - approver
            - admin
            - owner
            - auditor
          description: Role assigned to the entity within the group.
          example: "approver"
    EntityMembershipRemove:
      type: object
      required:
        - entity
      properties:
        entity:
          $ref: "#/components/schemas/EntityReference"
...
    AddGroupEntitiesRequest:
      type: object
      description: Defines operations to add entities within a group
      required:
        - entities
      properties:
        entities:
          type: array
          items:
            $ref: "#/components/schemas/EntityMembershipAdd"
          description: List of entities to add to the group.
    RemoveGroupEntitiesRequest:
      type: object
      description: Defines operations to remove entities within a group
      required:
        - entities
      properties:
        entities:
          type: array
          items:
            $ref: "#/components/schemas/EntityMembershipRemove"
          description: List of entities to remove from the group.
    GroupMembership:
      type: object
      required:
        - entity
        - role
        - addedAt
      properties:
        entity:
          $ref: "#/components/schemas/EntityReference"
        role:
          type: string
          x-extensible-enum:
            - approver
            - admin
            - owner
            - auditor
          description: The role of the entity within the group.
          example: "approver"
        addedAt:
          type: string
          format: date-time
          description: Timestamp when the entity was added to the group.
          readOnly: true
          example: "2025-04-16T10:30:00Z"
...
    GroupCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Name of the group
        description:
          type: string
          description: Description of the group
    Group:
      type: object
      required:
        - id
        - name
        - entitiesCount
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the group
        name:
          type: string
          description: Name of the group
        description:
          type: string
          description: Description of the group
        entitiesCount:
          type: integer
          description: Number of entities in the group
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    Pagination:
      type: object
      required:
        - total
        - page
        - limit
      properties:
        total:
          type: integer
          description: Total number of items
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Items per page
    APIError:
      type: object
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Error message
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []